+++ build.sbt
@@ -187,7 +187,6 @@ lazy val models = (project in file("models"))
     includeFilter in PB.generate := new SimpleFileFilter(
       protobufSubDirectoryFilter(
         "io/casperlabs/casper/consensus",
-        "io/casperlabs/ipc",
         "io/casperlabs/casper/protocol" // TODO: Eventually remove.
       )),
     PB.targets in Compile := Seq(
@@ -338,9 +337,18 @@ lazy val blockStorage = (project in file("block-storage"))
       catsCore,
       catsEffect,
       catsMtl
-    )
-  )
-  .dependsOn(shared, models % "compile->compile;test->test")
+    ),
+	  PB.protoSources in Compile := Seq(protobufDirectory),
+    includeFilter in PB.generate := new SimpleFileFilter(
+      protobufSubDirectoryFilter(
+        "io/casperlabs/storage"
+      )),
+    PB.targets in Compile := Seq(
+      scalapb.gen(flatPackage = true) -> (sourceManaged in Compile).value,
+      grpcmonix.generators.GrpcMonixGenerator(flatPackage = true) -> (sourceManaged in Compile).value
+  	)
+	)
+  .dependsOn(shared,smartContracts,models % "compile->compile;test->test")
 
 // Smart contract execution.
 lazy val smartContracts = (project in file("smart-contracts"))
@@ -353,6 +361,15 @@ lazy val smartContracts = (project in file("smart-contracts"))
       grpcNetty,
       nettyTransNativeEpoll,
       nettyTransNativeKqueue
+    ),
+    PB.protoSources in Compile := Seq(protobufDirectory),
+    includeFilter in PB.generate := new SimpleFileFilter(
+      protobufSubDirectoryFilter(
+        "io/casperlabs/ipc"
+      )),
+    PB.targets in Compile := Seq(
+      scalapb.gen(flatPackage = true) -> (sourceManaged in Compile).value,
+      grpcmonix.generators.GrpcMonixGenerator(flatPackage = true) -> (sourceManaged in Compile).value
     )
   )
   .dependsOn(shared, models)
