+++ casper/src/main/scala/io/casperlabs/casper/util/comm/ApproveBlockProtocol.scala
@@ -49,8 +49,7 @@ object ApproveBlockProtocol {
   def apply[F[_]](implicit instance: ApproveBlockProtocol[F]): ApproveBlockProtocol[F] = instance
 
   //For usage in tests only
-  def unsafe[
-      F[_]: Sync: ConnectionsCell: TransportLayer: Log: Time: Metrics: RPConfAsk: LastApprovedBlock](
+  def unsafe[F[_]: Sync: ConnectionsCell: TransportLayer: Log: Time: Metrics: RPConfAsk: LastApprovedBlock](
       block: BlockMessage,
       transforms: Seq[TransformEntry],
       trustedValidators: Set[ByteString],
@@ -71,8 +70,7 @@ object ApproveBlockProtocol {
       sigsF
     )
 
-  def of[
-      F[_]: Sync: ConnectionsCell: TransportLayer: Log: Time: Metrics: RPConfAsk: LastApprovedBlock](
+  def of[F[_]: Sync: ConnectionsCell: TransportLayer: Log: Time: Metrics: RPConfAsk: LastApprovedBlock](
       block: BlockMessage,
       transforms: Seq[TransformEntry],
       trustedValidators: Set[ByteString],
@@ -95,8 +93,7 @@ object ApproveBlockProtocol {
         sigsF
       )
 
-  private class ApproveBlockProtocolImpl[
-      F[_]: Sync: ConnectionsCell: TransportLayer: Log: Time: Metrics: RPConfAsk: LastApprovedBlock](
+  private class ApproveBlockProtocolImpl[F[_]: Sync: ConnectionsCell: TransportLayer: Log: Time: Metrics: RPConfAsk: LastApprovedBlock](
       val block: BlockMessage,
       val transforms: Seq[TransformEntry],
       val requiredSigs: Int,
