+++ casper/src/test/scala/io/casperlabs/casper/ManyValidatorsTest.scala
@@ -8,10 +8,11 @@ import io.casperlabs.casper.Estimator.BlockHash
 import io.casperlabs.casper.api.BlockAPI
 import io.casperlabs.casper.helper.BlockGenerator._
 import io.casperlabs.casper.helper._
-import io.casperlabs.casper.protocol.{BlockMessage, BlockMsgWithTransform, Bond}
+import io.casperlabs.casper.protocol.{BlockMessage, Bond}
 import io.casperlabs.metrics.Metrics.MetricsNOP
 import io.casperlabs.p2p.EffectsTestInstances.LogStub
 import io.casperlabs.shared.{Log, Time}
+import io.casperlabs.storage.BlockMsgWithTransform
 import monix.eval.Task
 import monix.execution.schedulers.CanBlock
 import org.scalatest.{FlatSpec, Matchers}
