+++ casper/src/test/scala/io/casperlabs/casper/ValidateTest.scala
@@ -21,6 +21,7 @@ import io.casperlabs.casper.util.execengine.{ExecEngineUtil, ExecutionEngineServ
 import io.casperlabs.ipc.TransformEntry
 import io.casperlabs.models.BlockMetadata
 import io.casperlabs.smartcontracts.ExecutionEngineService
+import io.casperlabs.storage.BlockMsgWithTransform
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
