+++ casper/src/test/scala/io/casperlabs/casper/api/BlockQueryResponseAPITest.scala
@@ -14,6 +14,7 @@ import io.casperlabs.casper.util.ProtoUtil
 import io.casperlabs.p2p.EffectsTestInstances.{LogStub, LogicalTime}
 import org.scalatest.{FlatSpec, Matchers}
 import io.casperlabs.catscontrib.TaskContrib._
+import io.casperlabs.storage.BlockMsgWithTransform
 import monix.eval.Task
 
 import scala.collection.immutable.HashMap
