+++ casper/src/test/scala/io/casperlabs/casper/api/BlocksResponseAPITest.scala
@@ -11,6 +11,7 @@ import io.casperlabs.casper.helper._
 import io.casperlabs.casper.helper.BlockGenerator._
 import io.casperlabs.casper.helper.BlockUtil.generateValidator
 import io.casperlabs.p2p.EffectsTestInstances.LogStub
+import io.casperlabs.storage.BlockMsgWithTransform
 import monix.eval.Task
 import org.scalatest.{FlatSpec, Matchers}
